refactor: Move logic to utils
